Make SendZeroCopy independent of libzmq version
SendZeroCopy can be now places into the general zmq sources since the detection
of libzmq version is happening at compile time.
