+++ zmq.go
@@ -22,6 +22,7 @@ package gozmq
 #include <zmq.h>
 #include <stdlib.h>
 #include <string.h>
+#include "zc.h"
 */
 import "C"
 
@@ -370,6 +371,45 @@ func (s *Socket) SendMultipart(parts [][]byte, flags SendRecvOption) (err error)
 	}
 	err = s.Send(parts[(len(parts)-1)], flags)
 	return
+}
+
+// Send a message to the socket using 0MQ zero copy.
+// SendZeroCopy is suitable only for large messages since it (for now)
+// incorporates some synchronization overhead in its Go part.
+func (s *Socket) SendZeroCopy(data []byte, flags SendRecvOption) error {
+	rc, err := C.gozmq_zc_send(s.apiSocket(),
+		unsafe.Pointer(&data[0]), C.size_t(len(data)), C.int(flags))
+	if rc == -1 {
+		return casterr(err)
+	}
+	return nil
+}
+
+// Global state necessary for the zero copy cleanup callback.
+// FIXME: Try to get rid of this global state.
+var (
+	zcSeq     int
+	zcSeqLock sync.Mutex
+
+	zcData     map[int][]byte = make(map[int][]byte)
+	zcDataLock sync.Mutex
+)
+
+//export gozmq_zc_seq
+func gozmq_zc_seq() C.int {
+	zcSeqLock.Lock()
+	defer zcSeqLock.Unlock()
+
+	zcSeq++
+	return C.int(zcSeq)
+}
+
+//export gozmq_zc_free_msg
+func gozmq_zc_free_msg(seq C.int) {
+	zcDataLock.Lock()
+	defer zcDataLock.Unlock()
+
+	delete(zcData, int(seq))
 }
 
 // Receive a multipart message.
